+++ INTEL/Algorithm/Problems/ProblemFactory.cs
@@ -9,7 +9,7 @@ namespace INTEL
     abstract class ProblemFactory
     {
         public abstract string Name { get; }
-        public decimal MaxFitness { get; protected set; }
+        public double MaxFitness { get; protected set; }
 
         public abstract void Initialize();
         public abstract Problem[] Create();
